[refactor] 게시글 목록 페이지 템플릿으로 변경
기존의 bulletin board로 한 뎁쓰 나눈게 관리가 어려워
template으로 변경
