+++ src/components/pages/ArticleList/index.tsx
@@ -1,16 +1,47 @@
 import React from "react";
-import { ArticleListProvider } from "~/contexts/ArticleList";
-import { CategoryProvider } from "~/contexts/Category";
-import View from "./view";
+import ArticleListView from "~/components/template/ArticleList";
+import { CategoryProvider, useCategory } from "~/contexts/Category";
+import { useAuth } from "~/contexts/Auth";
+import {
+    ArticleListProvider,
+    useArticleList,
+    Action as ArticleListAction,
+} from "~/contexts/ArticleList";
 
-const ArticleList = () => {
+const View = () => {
+    const {
+        state: { notices, articles, articlesCount },
+        dispatch: articleListDispatch,
+    } = useArticleList();
+    const {
+        state: { categories },
+    } = useCategory();
+
+    const {
+        state: { user },
+    } = useAuth();
+
     return (
-        <CategoryProvider>
-            <ArticleListProvider>
-                <View />
-            </ArticleListProvider>
-        </CategoryProvider>
+        <ArticleListView
+            categories={categories}
+            user={user}
+            articles={articles}
+            notices={notices}
+            articlesCount={articlesCount}
+            onPageChange={({ selected: page }) => {
+                articleListDispatch({
+                    type: ArticleListAction.SET_PAGE,
+                    page,
+                });
+            }}
+        />
     );
 };
 
-export default ArticleList;
+export default () => (
+    <CategoryProvider>
+        <ArticleListProvider>
+            <View />
+        </ArticleListProvider>
+    </CategoryProvider>
+);
