--- src/components/pages/ArticleList/style.scss
@@ -1,8 +1,0 @@
-.body_wrapper {
-    display: flex;
-
-    & main {
-        flex: 1;
-        margin-bottom: auto;
-    }
-}
